HEEDLS-658 Refactor Group Courses actions into new controller
