+++ DigitalLearningSolutions.Web.Tests/ViewModels/TrackingSystem/Delegates/GroupCourses/GroupCoursesViewModelTests.cs
@@ -1,10 +1,11 @@
-﻿namespace DigitalLearningSolutions.Web.Tests.ViewModels.TrackingSystem.Delegates.DelegateGroups
+﻿namespace DigitalLearningSolutions.Web.Tests.ViewModels.TrackingSystem.Delegates.GroupCourses
 {
     using System.Linq;
     using DigitalLearningSolutions.Data.Models.DelegateGroups;
     using DigitalLearningSolutions.Data.Tests.NBuilderHelpers;
     using DigitalLearningSolutions.Web.Models.Enums;
     using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.DelegateGroups;
+    using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.GroupCourses;
     using FizzWare.NBuilder;
     using FluentAssertions;
     using FluentAssertions.Execution;
