+++ DigitalLearningSolutions.Web/Controllers/TrackingSystem/Delegates/DelegateGroupsController.cs
@@ -155,62 +155,7 @@
             return RedirectToAction("GroupDelegates", new { groupId });
         }
 
-        [Route("{groupId:int}/Courses/{page:int=1}")]
-        public IActionResult GroupCourses(int groupId, int page = 1)
-        {
-            var centreId = User.GetCentreId();
-            var groupName = groupsService.GetGroupName(groupId, centreId);
-
-            if (groupName == null)
-            {
-                return NotFound();
-            }
-
-            var categoryIdFilter = User.GetAdminCourseCategoryFilter();
-
-            var groupCourses = groupsService.GetGroupCoursesForCategory(groupId, centreId, categoryIdFilter);
-
-            var model = new GroupCoursesViewModel(groupId, groupName, groupCourses, page);
-
-            return View(model);
-        }
-
-        [Route("{groupId:int}/Courses/{groupCustomisationId:int}/Remove")]
-        [ServiceFilter(typeof(VerifyAdminUserCanAccessGroup))]
-        [ServiceFilter(typeof(VerifyAdminUserCanAccessGroupCourse))]
-        public IActionResult RemoveGroupCourse(int groupId, int groupCustomisationId)
-        {
-            var centreId = User.GetCentreId();
-            var groupName = groupsService.GetGroupName(groupId, centreId);
-            var groupCourse = groupsService.GetUsableGroupCourseForCentre(groupCustomisationId, groupId, centreId);
-
-            var model = new RemoveGroupCourseViewModel(
-                groupCourse!.GroupCustomisationId,
-                groupCourse.CourseName,
-                groupName!
-            );
-
-            return View(model);
-        }
-
-        [HttpPost("{groupId:int}/Courses/{groupCustomisationId:int}/Remove")]
-        [ServiceFilter(typeof(VerifyAdminUserCanAccessGroup))]
-        [ServiceFilter(typeof(VerifyAdminUserCanAccessGroupCourse))]
-        public IActionResult RemoveGroupCourse(int groupId, int groupCustomisationId, RemoveGroupCourseViewModel model)
-        {
-            if (!ModelState.IsValid)
-            {
-                return View(model);
-            }
-
-            groupsService.RemoveGroupCourseAndRelatedProgress(
-                groupCustomisationId,
-                groupId,
-                model.DeleteStartedEnrolments
-            );
-
-            return RedirectToAction(nameof(GroupCourses), new { groupId });
-        }
+        
 
         [Route("{groupId:int}/Delete")]
         [ServiceFilter(typeof(VerifyAdminUserCanAccessGroup))]
@@ -364,57 +309,6 @@
             );
 
             return RedirectToAction("Index");
-        }
-
-        [HttpGet]
-        [Route("{groupId:int}/Courses/Add/SelectCourse")]
-        [ServiceFilter(typeof(VerifyAdminUserCanAccessGroup))]
-        public IActionResult AddCourseToGroupSelectCourse(int groupId)
-        {
-            var centreId = User.GetCentreId();
-
-            var adminCategoryFilter = User.GetAdminCourseCategoryFilter();
-
-            var courses = courseService.GetEligibleCoursesToAddToGroup(centreId, adminCategoryFilter, groupId);
-
-            var groupName = groupsService.GetGroupName(groupId, centreId);
-
-            var model = new AddCourseToGroupCoursesViewModel(courses, groupId, groupName!);
-
-            return View(model);
-        }
-
-        [HttpGet]
-        [Route("{groupId:int}/Courses/Add/{customisationId:int}")]
-        [ServiceFilter(typeof(VerifyAdminUserCanAccessGroup))]
-        [ServiceFilter(typeof(VerifyAdminUserCanViewCourse))]
-        public IActionResult AddCourseToGroup(int groupId, int customisationId)
-        {
-            var centreId = User.GetCentreId();
-            var groupLabel = groupsService.GetGroupName(groupId, centreId)!;
-            var courseCategoryId = courseService.GetCourseCategoryId(customisationId, centreId)!.Value;
-            var courseNameInfo = courseService.GetCourseNameAndApplication(customisationId)!;
-            var supervisors = userService.GetSupervisorsAtCentreForCategory(centreId, courseCategoryId);
-            var viewModel = new AddCourseViewModel(groupId, customisationId, supervisors, groupLabel, courseNameInfo);
-            return View(viewModel);
-        }
-
-        [HttpPost]
-        [Route("{groupId:int}/Courses/Add/{customisationId:int}")]
-        [ServiceFilter(typeof(VerifyAdminUserCanAccessGroup))]
-        [ServiceFilter(typeof(VerifyAdminUserCanViewCourse))]
-        public IActionResult AddCourseToGroup(AddCourseFormData formData, int groupId, int customisationId)
-        {
-            if (!ModelState.IsValid)
-            {
-                var courseCategoryId = courseService.GetCourseCategoryId(customisationId, User.GetCentreId())!.Value;
-                var supervisors = userService.GetSupervisorsAtCentreForCategory(User.GetCentreId(), courseCategoryId);
-                var model = new AddCourseViewModel(formData, groupId, customisationId, supervisors);
-                return View(model);
-            }
-
-            // TODO HEEDLS-658 Save + Confirmation page
-            return RedirectToAction("GroupCourses", new { groupId });
         }
 
         private IEnumerable<CustomPrompt> GetRegistrationPromptsWithSetOptions(int centreId)
