+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/GroupCourses/GroupCoursesViewModel.cs
@@ -1,10 +1,11 @@
-﻿namespace DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.DelegateGroups
+﻿namespace DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.GroupCourses
 {
     using System.Collections.Generic;
     using System.Linq;
     using DigitalLearningSolutions.Data.Models.DelegateGroups;
     using DigitalLearningSolutions.Web.Models.Enums;
     using DigitalLearningSolutions.Web.ViewModels.Common.SearchablePage;
+    using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.DelegateGroups;
 
     public class GroupCoursesViewModel : BasePaginatedViewModel
     {
