+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/DelegateGroups/_SearchableDelegateGroupCard.cshtml
@@ -110,8 +110,8 @@
 
       <a class="nhsuk-button expander-card__button"
          role="button"
-         asp-controller="DelegateGroups"
-         asp-action="GroupCourses"
+         asp-controller="GroupCourses"
+         asp-action="Index"
          asp-route-groupId="@Model.Id"
          asp-route-page="1">
         View courses
