+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/GroupCourses/AddCourseToGroup.cshtml
@@ -1,4 +1,4 @@
-﻿@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.DelegateGroups
+﻿@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.GroupCourses
 @model AddCourseViewModel
 
 @{
@@ -42,10 +42,10 @@
                         hint-text=""
                         css-class="nhsuk-u-width-one-half" />
 
-      <a class="nhsuk-button nhsuk-button--secondary" asp-controller="DelegateGroups" asp-action="AddCourseToGroupSelectCourse" asp-route-groupId="@Model.GroupId">Back</a>
+      <a class="nhsuk-button nhsuk-button--secondary" asp-controller="GroupCourses" asp-action="AddCourseToGroupSelectCourse" asp-route-groupId="@Model.GroupId">Back</a>
       <button class="nhsuk-button" type="submit">Add course</button>
     </form>
 
-    <vc:cancel-link asp-controller="DelegateGroups" asp-action="GroupCourses" asp-all-route-data="@routeParamsForBackLink" />
+    <vc:cancel-link asp-controller="GroupCourses" asp-action="Index" asp-all-route-data="@routeParamsForBackLink" />
   </div>
 </div>
