+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/GroupCourses/RemoveGroupCourse.cshtml
@@ -1,5 +1,5 @@
 ﻿@inject IConfiguration Configuration;
-@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.DelegateGroups
+@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.GroupCourses
 @using Microsoft.Extensions.Configuration
 @model RemoveGroupCourseViewModel
 
@@ -8,7 +8,7 @@
   ViewData["Title"] = $"{(errorHasOccurred ? "Error: " : "")}Groups - Remove group course";
 
   var cancelLinkRouteData = new Dictionary<string, string> {
-    { "groupId", ViewContext.RouteData.Values["groupId"].ToString()! }
+    { "groupId", ViewContext.RouteData.Values["groupId"].ToString()! },
   };
 }
 
@@ -58,6 +58,6 @@
       <button class="nhsuk-button delete-button" type="submit">Remove course</button>
     </form>
 
-    <vc:cancel-link asp-controller="DelegateGroups" asp-action="GroupCourses" asp-all-route-data="@cancelLinkRouteData" />
+    <vc:cancel-link asp-controller="GroupCourses" asp-action="Index" asp-all-route-data="@cancelLinkRouteData" />
   </div>
 </div>
